BLUEs

Documentation for BLUEs.

BLUEs.Estimate

BLUEs.OverdeterminedProblem

BLUEs.OverdeterminedProblem

BLUEs.UnderdeterminedProblem

BLUEs.UnderdeterminedProblem

BLUEs.combine

BLUEs.controlcost

BLUEs.cost

BLUEs.datacost

BLUEs.flipped_mult

BLUEs.gradient

BLUEs.hessian

BLUEs.impulseresponse

BLUEs.misfitgradient

BLUEs.rmscontrol

BLUEs.rmserror

BLUEs.solve

BLUEs.solve

BLUEs.solve_hessian

BLUEs.solve_textbook

BLUEs.symmetric_innerproduct

Base.:*

Base.:*

Base.:+

Base.:\

Base.getproperty

LinearAlgebra.pinv

BLUEs.Estimate — Type

struct Estimate{Tv <: Number, TP <: Number, Nv, NP}

a structure with some vector of values x and associated uncertainty matrix P

Fields

v :: AbstractArray{T, Nv}

P :: AbstractArray{T, NP}

BLUEs.OverdeterminedProblem — Type

struct OverdeterminedProblem

a structure (NamedTuple version) with fields

- `y::Union{<: AbstractVector,NamedTuple}`: "observations", namedtuples of vectors
- `E :: Union{<: AbstractMatrix, NamedTuple}`: model matrices 
- `Cnn⁻¹ :: Union{<: AbstractMatrix, NamedTuple}`: namedtuple of inverse noise covariance matrix
- `Cxx⁻¹ :: Union{<: AbstractMatrix, Missing}`: tapering matrices, NOT namedtuples 
- `x₀ :: Union{<: AbstractVector, Missing}`: first guess vector

BLUEs.OverdeterminedProblem — Method

function OverdeterminedProblem

generates OverdeterminedProblem structure with x₀ = missing, Cxx = missing

BLUEs.UnderdeterminedProblem — Type

struct UnderdeterminedProblem

a structure with fields

- `y::AbstractVector`: vector of "observations"
- `E::AbstractMatrix`: model matrix 
- `Cnn::AbstractMatrix`: noise covariance matrix 
- `Cxx::Union{AbstractMatrix}`: tapering matrix 
- `x₀::Union{AbstractVector, Missing}`: first guess vector

BLUEs.UnderdeterminedProblem — Method

function UnderdeterminedProblem

generates UnderdeterminedProblem structure with x₀ = missing, still requires Cxx

BLUEs.combine — Method

combine(x0::Estimate,y::Estimate,f::Function)

Arguments

x0::Estimate: estimate 1

y::Estimate: estimate 2

f::Function: function that relates f(x0) = y

Returns

xtilde::Estimate: combined estimate

\[{\bf E}_i (\tau) = \frac{1}{N} \int_{t - \tau}^{t} {\bf G}'^{\dagger} (t^* + \tau - t) ~ {\bf D}_i ~ {\bf G}' (t - t^*) ~ d t ^* , \]

BLUEs.controlcost — Method

Cost function contribution from control vector

BLUEs.cost — Method

Compute cost function

BLUEs.datacost — Method

Cost function contribution from observations

BLUEs.flipped_mult — Method

function flipped_mult

multiply in opposite order given, needs to be defined for impulseresponse

BLUEs.gradient — Method

function gradient

compute ∂J∂x = -(Eᵀ∂J∂n) = -2(Eᵀ(Cnn⁻¹y)) or -2Eᵀ[Cnn⁻¹(y - Ex₀)]

BLUEs.hessian — Method

function hessian

compute Eᵀ(Cnn⁻¹E) or Eᵀ(Cnn⁻¹E) + Cxx⁻¹
depending on if prior is available

BLUEs.impulseresponse — Method

function impulseresponse(x₀,M)

Probe a function to determine its linear response in matrix form.
Assumes units are needed and available.
A simpler function to handle cases without units would be nice.

funk:: function to be probed
x:: input variable
args:: the arguments that follow x in `funk`

BLUEs.misfitgradient — Method

function misfitgradient

returns 2(Cnn⁻¹y) or 2[Cnn⁻¹(y - Ex₀)]

BLUEs.rmscontrol — Function

function rmscontrol

compute (ũ-u₀)ᵀ(ũ-u₀), how much are we adjusting the control?
unweighted `controlcost`

# Args 
-`x̃`: Estimate
-`p`: problem
-`dim3`: allows to access third dimension, which is assumed to be state var. dim.

BLUEs.rmserror — Method

function rmserror

compute nᵀn, how closely are we fitting the data?
unweighted `datacost`

BLUEs.solve — Method

function solve

    Solve overdetermined problem

    optional alg= :textbook or :hessian

BLUEs.solve — Method

function solve

    Solve underdetermined problem

BLUEs.solve_hessian — Method

function solve_hessian

Solving y = Ex

H = Eᵀ(Cnn⁻¹E) + Cxx⁻¹ ∂J/∂x = -E^T ∂J/∂n ∂J/∂n = 2n n = y - Ex₀ x̃ = -1/2 H⁻¹ ∂J/∂x

BLUEs.solve_textbook — Method

function solve_textbook

Solves overdetermined problem
y = Ex with associated uncertainty Cnn⁻¹ that is used for weighting
       optionally, can have prior information as well 

x̃ = (EᵀCnn⁻¹E)⁻¹[(Cnn⁻¹E)ᵀy]
Cx̃x̃ = (EᵀCnn⁻¹E)⁻¹

If prior information (Cxx⁻¹, x₀) is available
x̃ = (EᵀCnn⁻¹E + Cxx⁻¹)⁻¹[(Cnn⁻¹E)ᵀy + Cxx⁻¹x₀]
Cx̃x̃ = (EᵀCnn⁻¹E)⁻¹

See equations 1.208/1.209 in Dynamical Insights from Data

BLUEs.symmetric_innerproduct — Method

for NamedTuple, add up each Hessian contribution

Base.:* — Method

Matrix multiplication for Estimate includes
error propagation.

Base.:* — Method

multiplication for `NamedTuple`s

Matrix multiply, M*x

Base.:+ — Method

Matrix addition for Estimate includes
error propagation. Follow pp. 95, Sec. 2.5.5,
Recursive Least Squares, "Dynamical Insights from Data" class notes

Base.:\ — Method

Matrix left divide for Estimate includes
error propagation.

Base.getproperty — Method

function getproperty(x::Estimate, d::Symbol)

Fields of Estimate

x::Vector{Measurement}: Estimate and 1σ uncertainty

err::Vector{Number}: 1σ uncertainty

val::Vector{Number}: central value of estimate

P::Matrix{Number}: estimate uncertainty matrix

LinearAlgebra.pinv — Method

function pinv

Left pseudo-inverse (i.e., least-squares estimator)