BLUEs

Documentation for BLUEs.

BLUEs.DimEstimate

BLUEs.Estimate

BLUEs.OverdeterminedProblem

BLUEs.OverdeterminedProblem

BLUEs.UnderdeterminedProblem

BLUEs.UnderdeterminedProblem

BLUEs.controlcost

BLUEs.convolve

BLUEs.convolve

BLUEs.cost

BLUEs.datacost

BLUEs.flipped_mult

BLUEs.gradient

BLUEs.hessian

BLUEs.impulseresponse

BLUEs.misfitgradient

BLUEs.predictobs

BLUEs.solve

BLUEs.solve

BLUEs.solve_hessian

BLUEs.solve_textbook

BLUEs.symmetric_innerproduct

Base.:*

Base.:*

Base.:+

Base.getproperty

LinearAlgebra.pinv

BLUEs.DimEstimate — Type

struct DimEstimate{Tv <: Number,TC <: Number} 

A structure with some vector of values v and associated covariance matrix C.

Differs from Estimate in that axis dimensions are added.

Fields

v :: AbstractVector{Tv}

C :: AbstractMatrix{TC}

dims :: Tuple

BLUEs.Estimate — Type

struct Estimate{Tv <: Number,TC <: Number}

a structure with some vector of values v and associated covariance matrix C

Fields

v :: AbstractVector{Tv}

C :: AbstractMatrix{TC}

BLUEs.OverdeterminedProblem — Type

struct OverdeterminedProblem

a structure (NamedTuple version) with fields

- `y::Union{<: AbstractVector,NamedTuple}`: "observations", namedtuples of vectors
- `E :: Union{<: AbstractMatrix, NamedTuple}`: model matrices 
- `Cnn⁻¹ :: Union{<: AbstractMatrix, NamedTuple}`: namedtuple of inverse noise covariance matrix
- `Cxx⁻¹ :: Union{<: AbstractMatrix, Missing}`: tapering matrices, NOT namedtuples 
- `x₀ :: Union{<: AbstractVector, Missing}`: first guess vector

BLUEs.OverdeterminedProblem — Method

function OverdeterminedProblem

generates OverdeterminedProblem structure with x₀ = missing, Cxx = missing

BLUEs.UnderdeterminedProblem — Type

struct UnderdeterminedProblem

a structure with fields

- `y::AbstractVector`: vector of "observations"
- `E::AbstractMatrix`: model matrix 
- `Cnn::AbstractMatrix`: noise covariance matrix 
- `Cxx::Union{AbstractMatrix}`: tapering matrix 
- `x₀::Union{AbstractVector, Missing}`: first guess vector

BLUEs.UnderdeterminedProblem — Method

function UnderdeterminedProblem

generates UnderdeterminedProblem structure with x₀ = missing, still requires Cxx

BLUEs.controlcost — Method

Cost function contribution from control vector

BLUEs.convolve — Method

function convolve(x::AbstractDimArray, E::AbstractDimArray, coeffs::UnitfulMatrix} the coeffs argument signifies that x is a 3D array (i.e. >1 state variables)

this function both convolves, and linearly combines the propagated state variables

BLUEs.convolve — Method

function convolve(E::AbstractDimArray,x::AbstractDimArray)

Take the convolution of E and x
Account for proper overlap of dimensions
Sum and take into account units.

BLUEs.cost — Method

Compute cost function

BLUEs.datacost — Method

Cost function contribution from observations

BLUEs.flipped_mult — Method

function flipped_mult

multiply in opposite order given, needs to be defined for impulseresponse

BLUEs.gradient — Method

function gradient

compute ∂J∂x = -(Eᵀ∂J∂n) = -2(Eᵀ(Cnn⁻¹y)) or -2Eᵀ[Cnn⁻¹(y - Ex₀)]

BLUEs.hessian — Method

function hessian

compute Eᵀ(Cnn⁻¹E) or Eᵀ(Cnn⁻¹E) + Cxx⁻¹
depending on if prior is available

BLUEs.impulseresponse — Method

function impulseresponse(x₀,M)

Probe a function to determine its linear response in matrix form.
Assumes units are needed and available.
A simpler function to handle cases without units would be nice.

funk:: function to be probed
x:: input variable
args:: the arguments that follow x in `funk`

BLUEs.misfitgradient — Method

function misfitgradient

returns 2(Cnn⁻¹y) or 2[Cnn⁻¹(y - Ex₀)]

BLUEs.predictobs — Method

function predictobs(funk,x...)

Get observations derived from function `funk`
y = funk(x...)
Turns out to not be useful so far.

BLUEs.solve — Method

function solve

    Solve overdetermined problem

    optional alg= :textbook or :hessian

BLUEs.solve — Method

function solve

    Solve underdetermined problem

BLUEs.solve_hessian — Method

function solve_hessian

Solving y = Ex

x̃ = -1/2 (Eᵀ(Cnn⁻¹E))⁻¹ -(Eᵀ 2(Cnn⁻¹y))
Cx̃x̃ = (Eᵀ(Cnn⁻¹E))⁻¹
(maybe???? doublecheck with Jake)

BLUEs.solve_textbook — Method

function solve_textbook

Solves overdetermined problem
y = Ex with associated uncertainty Cnn⁻¹ that is used for weighting
       optionally, can have prior information as well 

x̃ = (EᵀCnn⁻¹E)⁻¹[(Cnn⁻¹E)ᵀy]
Cx̃x̃ = (EᵀCnn⁻¹E)⁻¹

If prior information (Cxx⁻¹, x₀) is available
x̃ = (EᵀCnn⁻¹E + Cxx⁻¹)⁻¹[(Cnn⁻¹E)ᵀy + Cxx⁻¹x₀]
Cx̃x̃ = (EᵀCnn⁻¹E)⁻¹

See equations 1.208/1.209 in Dynamical Insights from Data

BLUEs.symmetric_innerproduct — Method

for NamedTuple, add up each Hessian contribution

Base.:* — Method

Matrix multiplication for Estimate includes
error propagation.

Base.:* — Method

multiplication for `NamedTuple`s

Base.:+ — Method

Matrix addition for Estimate includes
error propagation. Follow pp. 95, Sec. 2.5.5,
Recursive Least Squares, "Dynamical Insights from Data" class notes

Base.getproperty — Method

function getproperty(x::Estimate, d::Symbol)

Fields of Estimate

x::Vector{Measurement}: Estimate and 1σ uncertainty

σ::Vector{Number}: 1σ uncertainty

v::Vector{Number}: central value of estimate

C::Matrix{Number}: estimate uncertainty matrix

LinearAlgebra.pinv — Method

function pinv

Left pseudo-inverse (i.e., least-squares estimator)